+++ code_examples/train_atari/rnd_ppo/conf.yaml
@@ -37,4 +37,4 @@ ext_adv_coeff: 2.0
 int_adv_coeff: 1.0
 predictor_proportion: 0.25
 pre_normalization_steps: 50
-wandb_key: b5ef9811ad92fac918facca483c7b6caa73df290
+wandb_key: null
